test: move uptime executor tests into feature module
